Used lombok for getters and setters
