--- src/main/java/auditLogEntry/AuditLogEntry.java
@@ -1,54 +1,0 @@
-package auditLogEntry;
-
-/**
- * encapsulates details about single audit log record
- * An instance of this class is used to log operation details into database
- */
-
-public class AuditLogEntry {
-    private String inputFilePath;
-    private String searchWord;
-    private String result;
-    private int count;
-    private String errorMessage;
-
-    public String getInputFilePath() {
-        return inputFilePath;
-    }
-
-    public void setInputFilePath(String inputFilePath) {
-        this.inputFilePath = inputFilePath;
-    }
-
-    public String getSearchWord() {
-        return searchWord;
-    }
-
-    public void setSearchWord(String searchWord) {
-        this.searchWord = searchWord;
-    }
-
-    public String getResult() {
-        return result;
-    }
-
-    public void setResult(String result) {
-        this.result = result;
-    }
-
-    public int getCount() {
-        return count;
-    }
-
-    public void setCount(int count) {
-        this.count = count;
-    }
-
-    public String getErrorMessage() {
-        return errorMessage;
-    }
-
-    public void setErrorMessage(String errorMessage) {
-        this.errorMessage = errorMessage;
-    }
-}
